// app/lib/types/audit.db.types.ts
export enum SecurityAuditAction {
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAILURE = "LOGIN_FAILURE",
  REGISTRATION = "REGISTRATION",
  LOGOUT = "LOGOUT",
  PASSWORD_CHANGE = "PASSWORD_CHANGE",
  PASSWORD_RESET = "PASSWORD_RESET",
  FORCE_PASSWORD_RESET = "FORCE_PASSWORD_RESET",
  PASSWORD_RESET_REQUEST = "PASSWORD_RESET_REQUEST",
  ACCOUNT_LOCKED = "ACCOUNT_LOCKED",
  ACCOUNT_UNLOCKED = "ACCOUNT_UNLOCKED",
  EMAIL_CHANGE_REQUEST = "EMAIL_CHANGE_REQUEST",
  EMAIL_CHANGE = "EMAIL_CHANGE",
  EMAIL_CHANGE_CONFIRMATION = "EMAIL_CHANGE_CONFIRMATION",
  EMAIL_CHANGE_FAILURE = "EMAIL_CHANGE_FAILURE",
  EMAIL_VERIFICATION = "EMAIL_VERIFICATION",
  TWO_FACTOR_AUTH_REQUIRED = "TWO_FACTOR_AUTH_REQUIRED",
  TWO_FACTOR_AUTH_SUCCESS = "TWO_FACTOR_AUTH_SUCCESS",
  SESSION_CREATED = "SESSION_CREATED",
  REVOKE_ALL_SESSIONS = "REVOKE_ALL_SESSIONS",
  SESSION_REVOKED = "SESSION_REVOKED",
  SESSION_TERMINATED = "SESSION_TERMINATED",
  BACKUP_CODE_RECOVERY = "BACKUP_CODE_RECOVERY",
  BACKUP_CODE_REGENERATED = "BACKUP_CODE_REGENERATED",
  BACKUP_CODE_USED = "BACKUP_CODE_USED",
  VERIFICATION_EMAIL_REQUEST = "VERIFICATION_EMAIL_REQUEST",
  VERIFICATION_EMAIL_SENT = "VERIFICATION_EMAIL_SENT",
  VERIFICATION_EMAIL_CONFIRMED = "VERIFICATION_EMAIL_CONFIRMED",
  VERIFICATION_EMAIL_RESENT = "VERIFICATION_EMAIL_RESENT",
  VERIFICATION_EMAIL_FAILED = "VERIFICATION_EMAIL_FAILED",
  VERIFICATION_EMAIL_CODE_RESENT = "VERIFICATION_EMAIL_CODE_RESENT",
  VERIFICATION_EMAIL_CODE_FAILED = "VERIFICATION_EMAIL_CODE_FAILED",
  VERIFICATION_EMAIL_CODE_USED = "VERIFICATION_EMAIL_CODE_USED",
  VERIFICATION_EMAIL_CODE_EXPIRED = "VERIFICATION_EMAIL_CODE_EXPIRED",
  VERIFICATION_EMAIL_CODE_INVALID = "VERIFICATION_EMAIL_CODE_INVALID",
  VERIFICATION_EMAIL_CODE_RECOVERED = "VERIFICATION_EMAIL_CODE_RECOVERED",
  VERIFICATION_EMAIL_CODE_REGENERATED = "VERIFICATION_EMAIL_CODE_REGENERATED",
  VERIFICATION_EMAIL_CODE_RECOVERY_FAILED = "VERIFICATION_EMAIL_CODE_RECOVERY_FAILED",
  VERIFICATION_EMAIL_CODE_RECOVERY_REQUEST = "VERIFICATION_EMAIL_CODE_RECOVERY_REQUEST",
  VERIFICATION_EMAIL_CODE_RECOVERY_SENT = "VERIFICATION_EMAIL_CODE_RECOVERY_SENT",
  VERIFICATION_EMAIL_FAILURE = "VERIFICATION_EMAIL_FAILURE",
  USER_UPDATE = "USER_UPDATE",
  USER_DELETE = "USER_DELETE",
  SECURITY_ACTION = "SECURITY_ACTION",
  AUDIT_LOG_EXPORT = "AUDIT_LOG_EXPORT",
  AUDIT_LOG_DELETE = "AUDIT_LOG_DELETE",
  AUDIT_LOG_CLEAR = "AUDIT_LOG_CLEAR",
  AUDIT_LOG_VIEW = "AUDIT_LOG_VIEW",
  NEW_DEVICE_ALERT_SENT = "NEW_DEVICE_ALERT_SENT",
  IMPOSSIBLE_TRAVEL_ALERT_SENT = "IMPOSSIBLE_TRAVEL_ALERT_SENT",
  SUSPICIOUS_LOGIN = "SUSPICIOUS_LOGIN",
  PASSWORD_CHANGED = "PASSWORD_CHANGED",
  NEW_DEVICE = "NEW_DEVICE",
  IMPOSSIBLE_TRAVEL = "IMPOSSIBLE_TRAVEL",
  SECURITY_SETTINGS_CHANGED = "SECURITY_SETTINGS_CHANGED",
  BACKUP_CODES_GENERATED = "BACKUP_CODES_GENERATED",
  BACKUP_CODES_CONSUMED = "BACKUP_CODES_CONSUMED",
  MFA_ENABLED = "MFA_ENABLED",
  MFA_DISABLED = "MFA_DISABLED",
  BOT_DETECTED = "BOT_DETECTED",
  DATA_EXPORT_REQUESTED = "DATA_EXPORT_REQUESTED",
  ACCOUNT_RECOVERY_INITIATED = "ACCOUNT_RECOVERY_INITIATED",
  UNUSUAL_ACTIVITY = "UNUSUAL_ACTIVITY",
  CRITICAL_ALERT = "CRITICAL_ALERT",
  LOW_BACKUP_CODES = "LOW_BACKUP_CODES",
  SUSPICIOUS_ACTIVITY = "SUSPICIOUS_ACTIVITY",
}
// ...
export interface AuditLogDetails {
  success: boolean;
  device: DeviceInfo;
  message: string;
  // [key: string]: any; // Additional dynamic details
}